requirements.txt
attrdict @ git+https://github.com/lucasmccabe/AttrDict.git
pyyaml==6.0